# PERF: Rolling/Expanding.cov/corr