--- a/pandas/core/window/rolling.py
+++ b/pandas/core/window/rolling.py
@@ -835,11 +835,37 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
         return result
 
     def _apply_pairwise(self, target, other, pairwise, func):
-        # Need to manually drop the grouping column first
+        # Manually drop the grouping column first
         target = target.drop(columns=list(self._groupby.keys), errors="ignore")
         result = super()._apply_pairwise(target, other, pairwise, func)
-        # Modify the resulting index to include the groupby level
 
+        # Modify the resulting index to include the groupby level
+        groupby_codes = self._groupby.grouper.codes
+        groupby_levels = self._groupby.grouper.levels
+
+        group_indices = self._groupby.grouper.indices.values()
+        if group_indices:
+            indexer = np.concatenate(list(group_indices))
+        else:
+            indexer = np.array([], dtype=np.intp)
+
+        if target.ndim == 1:
+            num_target_columns = 1
+        else:
+            num_target_columns = len(target.columns)
+        groupby_codes = [
+            np.repeat(c.take(indexer), num_target_columns) for c in groupby_codes
+        ]
+
+        result_codes = groupby_codes + list(result.index.codes)
+        result_levels = groupby_levels + list(result.index.levels)
+        result_names = list(self._groupby.keys) + list(result.index.names)
+
+        result_index = MultiIndex(
+            result_levels, result_codes, names=result_names, verify_integrity=False
+        )
+
+        result.index = result_index
         return result
 
     def _create_data(self, obj: FrameOrSeries) -> FrameOrSeries: