--- a/pandas/core/window/rolling.py
+++ b/pandas/core/window/rolling.py
@@ -51,6 +51,7 @@ from pandas.core.dtypes.missing import notna
 from pandas.core.algorithms import factorize
 from pandas.core.apply import ResamplerWindowApply
 from pandas.core.base import DataError, SelectionMixin
+import pandas.core.common as common
 from pandas.core.construction import extract_array
 from pandas.core.groupby.base import GotItemMixin, ShallowMixin
 from pandas.core.indexes.api import Index, MultiIndex
@@ -819,6 +820,7 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
 
         # if the index of the original dataframe needs to be preserved, append
         # this index (but reordered) to the codes/levels from the groupby
+        breakpoint()
         if grouped_object_index is not None:
             idx = grouped_object_index.take(indexer)
             if not isinstance(idx, MultiIndex):
@@ -834,11 +836,8 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
         return result
 
     def _apply_pairwise(self, target, other, pairwise, func):
-        def as_list(arg):
-            return [arg] if not isinstance(arg, (tuple, list)) else arg
-
         # Manually drop the grouping column first
-        target = target.drop(columns=as_list(self._groupby.keys), errors="ignore")
+        target = target.drop(columns=self._groupby.grouper.names, errors="ignore")
         result = super()._apply_pairwise(target, other, pairwise, func)
         # Modify the resulting index to include the groupby level
         if other is not None:
@@ -853,7 +852,9 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
                 ]
             )
 
-            gb_pairs = (as_list(pair) for pair in self._groupby.indices.keys())
+            gb_pairs = (
+                common.maybe_make_list(pair) for pair in self._groupby.indices.keys()
+            )
             groupby_codes = []
             groupby_levels = []
             # e.g. [[1, 2], [4, 5]] as [[1, 4], [2, 5]]
@@ -866,7 +867,7 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
             result_codes, result_levels = factorize(result.index)
             result_codes = groupby_codes + [result_codes]
             result_levels = groupby_levels + [result_levels]
-            result_names = as_list(self._groupby.keys) + [result.index.name]
+            result_names = self._groupby.grouper.names + [result.index.name]
 
         else:
             # When we evaluate the pairwise=True result, repeat the groupby
@@ -899,7 +900,7 @@ class BaseWindowGroupby(GotItemMixin, BaseWindow):
 
             result_codes = groupby_codes + result_codes
             result_levels = groupby_levels + result_levels
-            result_names = as_list(self._groupby.keys) + result_names
+            result_names = self._groupby.grouper.names + result_names
 
         result_index = MultiIndex(
             result_levels, result_codes, names=result_names, verify_integrity=False