… asv_bench/benchmarks/rolling.py
… doc/source/whatsnew/v1.3.0.rst
… pandas/tests/window/test_groupby.py